[REFACTOR] api 및 entity에 맞게 이름 변경
- ClassRoom => LectureRoom
- Professior => Teacher
