+++ src/components/templates/LectureRoomManagement/LectureRoomListView.js
@@ -1,34 +1,37 @@
 import React, { useEffect, useState } from 'react';
 import styled from 'styled-components';
-import RoundButton from '../../atoms/Button/RoundButton';
 import ContentHeader from '../../modules/ContentHeader/ContentHeader';
-import PreInfoViewContentMain from '../../modules/ContentMain/PreInfoViewContentMain';
+import PreInfoView from '../../modules/ContentMain/PreInfoView';
 
 const Wrapper = styled.div``;
 const head = [
   {
     key: 'lecture_room_id',
     name: '#',
+    type: 'id',
     width: '10%'
   },
   {
     key: 'building',
     name: '건물',
+    type: 'string',
     width: '30%'
   },
   {
     key: 'room_number',
     name: '호수',
+    type: 'number',
     width: '30%'
   },
   {
     key: 'capacity',
     name: '정원',
+    type: 'number',
     width: '30%'
   }
 ];
 
-const inActive = [
+const active = [
   {
     lecture_room_id: '1',
     building: 'D',
@@ -61,40 +64,22 @@ const inActive = [
   }
 ];
 
-const active = [
-  {
-    lecture_room_id: '6',
-    building: 'D',
-    room_number: 332,
-    capacity: 50
-  }
-];
-
-const ClassRoomListView = () => {
+const LectureRoomListView = () => {
   const headItem = head;
-  const [inActiveRows, setInActiveRows] = useState([]);
-  const [activeRows, setActiveRows] = useState([]);
+  const title = '강의실';
+  const headerTitle = `사전정보 - ${title}관리`;
+  const [rows, setRows] = useState([]);
 
   useEffect(() => {
-    setInActiveRows(inActive);
-    setActiveRows(active);
+    setRows(active);
   }, []);
 
   return (
     <Wrapper>
-      <ContentHeader title="강의실 관리">
-        <RoundButton>엑셀 업로드</RoundButton>
-        <RoundButton color="secondary">엑셀 다운로드</RoundButton>
-      </ContentHeader>
-      <PreInfoViewContentMain
-        head={headItem}
-        inActiveRows={inActiveRows}
-        activeRows={activeRows}
-        setInActiveRows={setInActiveRows}
-        setActiveRows={setActiveRows}
-      />
+      <ContentHeader title={headerTitle} />
+      <PreInfoView title={title} head={headItem} rows={rows} />
     </Wrapper>
   );
 };
 
-export default ClassRoomListView;
+export default LectureRoomListView;
